iw: Remove include of Makefile to avoid definition of invalid rules
This can be harmful for rules such as "clean" and implicit make rules
defined in iw's Makefile, such as "%.o", etc.

The problem was detected from this build warning:

  build/core/main.mk:1045: warning: overriding commands for target `clean'
  external/iw/Makefile:126: warning: ignoring old commands for target `clean'

Fix it by removing the include of Makefile in Android.mk.

For now, this creates even more duplication between the two files, in
the future we can try to prevent that by factoring out common code (such
as the list of *.o files, cflags, etc.) into a "common" file that can be
included in both makefiles.

TEST=Build iw from the AOSP tree and confirm it is built correctly.
  Confirmed that the warning about the "clean" rule is gone.

Change-Id: Ia927a71edbf0243dc0074a0ac6f6aa583ce636c5